refactor: 지하철 노선 정보 도메인 적용
